+++ api/routes/locations.ts
@@ -1,6 +1,6 @@
 import express, {NextFunction, Request, Response} from 'express';
 import mysqlDb from '../mysqlDb';
-import {Locations, LocationWithoutId} from '../types';
+import {Locations, LocationWithoutId, Product} from '../types';
 import {ResultSetHeader} from 'mysql2';
 
 const locationsRouter = express.Router();
@@ -8,7 +8,7 @@ const locationsRouter = express.Router();
 locationsRouter.get('/', async (req: Request, res: Response) => {
     const connection = mysqlDb.getConnection();
 
-    const [result] = await connection.query<Locations[]>('SELECT * FROM locations');
+    const [result] = await connection.query<Locations[]>('SELECT id, title FROM locations');
     res.send(result);
 });
 
@@ -29,13 +29,22 @@ locationsRouter.delete('/:id', async (req: Request, res: Response, next: NextFun
    const connection = mysqlDb.getConnection();
 
    try {
+       const [products] = await connection.query<Product[]>(
+            'SELECT id FROM products WHERE category_id = ?',
+            [id]
+        );
+
+       if (products.length > 0) {
+           return res.status(200).send({error: 'You cannot delete a location that contains products'});
+       }
+
        const [result] = await connection.query<ResultSetHeader>('DELETE FROM locations WHERE id = ?', [id]);
 
        if (result.affectedRows === 0) {
-           return res.status(404).send( 'Location not found');
-       }
+            return res.status(404).send('Category not found');
+        }
 
-        res.send({message: 'Location deleted successfully'});
+        res.send({message: 'Category deleted successfully'});
 
    } catch (e) {
        next(e);
@@ -71,7 +80,7 @@ locationsRouter.post('/', async (req: Request, res: Response, next: NextFunction
 
 locationsRouter.put('/:id', async (req: Request, res: Response, next: NextFunction) => {
     if (!req.body.title || !req.body.description) {
-        return res.status(400).send({error: 'Title is required'});
+        return res.status(400).send({error: 'Title and description are required'});
     }
 
     const id = req.params.id as string;
